requirements.txt
requests
beautifulsoup4
pandas
google-cloud-bigquery
pandas-gbq
re